+++ example/lib/src/file_picker_demo.dart
@@ -29,10 +29,18 @@ class _FilePickerDemoState extends State<FilePickerDemo> {
     try {
       if (_multiPick) {
         _path = null;
-        _paths = await FilePicker.getMultiFilePath(type: _pickingType, allowedExtensions: _extension?.replaceAll(' ', '')?.split(','));
+        _paths = await FilePicker.getMultiFilePath(
+            type: _pickingType,
+            allowedExtensions: (_extension?.isNotEmpty ?? false)
+                ? _extension?.replaceAll(' ', '')?.split(',')
+                : null);
       } else {
         _paths = null;
-        _path = await FilePicker.getFilePath(type: _pickingType, allowedExtensions: _extension?.replaceAll(' ', '')?.split(','));
+        _path = await FilePicker.getFilePath(
+            type: _pickingType,
+            allowedExtensions: (_extension?.isNotEmpty ?? false)
+                ? _extension?.replaceAll(' ', '')?.split(',')
+                : null);
       }
     } on PlatformException catch (e) {
       print("Unsupported operation" + e.toString());
@@ -40,7 +48,9 @@ class _FilePickerDemoState extends State<FilePickerDemo> {
     if (!mounted) return;
     setState(() {
       _loadingPath = false;
-      _fileName = _path != null ? _path.split('/').last : _paths != null ? _paths.keys.toString() : '...';
+      _fileName = _path != null
+          ? _path.split('/').last
+          : _paths != null ? _paths.keys.toString() : '...';
     });
   }
 
@@ -88,7 +98,7 @@ class _FilePickerDemoState extends State<FilePickerDemo> {
                       onChanged: (value) => setState(() {
                             _pickingType = value;
                             if (_pickingType != FileType.custom) {
-                              _controller.text = _extension = null;
+                              _controller.text = _extension = '';
                             }
                           })),
                 ),
@@ -99,7 +109,8 @@ class _FilePickerDemoState extends State<FilePickerDemo> {
                           maxLength: 15,
                           autovalidate: true,
                           controller: _controller,
-                          decoration: InputDecoration(labelText: 'File extension'),
+                          decoration:
+                              InputDecoration(labelText: 'File extension'),
                           keyboardType: TextInputType.text,
                           textCapitalization: TextCapitalization.none,
                         )
@@ -108,8 +119,10 @@ class _FilePickerDemoState extends State<FilePickerDemo> {
                 new ConstrainedBox(
                   constraints: BoxConstraints.tightFor(width: 200.0),
                   child: new SwitchListTile.adaptive(
-                    title: new Text('Pick multiple files', textAlign: TextAlign.right),
-                    onChanged: (bool value) => setState(() => _multiPick = value),
+                    title: new Text('Pick multiple files',
+                        textAlign: TextAlign.right),
+                    onChanged: (bool value) =>
+                        setState(() => _multiPick = value),
                     value: _multiPick,
                   ),
                 ),
@@ -122,18 +135,28 @@ class _FilePickerDemoState extends State<FilePickerDemo> {
                 ),
                 new Builder(
                   builder: (BuildContext context) => _loadingPath
-                      ? Padding(padding: const EdgeInsets.only(bottom: 10.0), child: const CircularProgressIndicator())
+                      ? Padding(
+                          padding: const EdgeInsets.only(bottom: 10.0),
+                          child: const CircularProgressIndicator())
                       : _path != null || _paths != null
                           ? new Container(
                               padding: const EdgeInsets.only(bottom: 30.0),
                               height: MediaQuery.of(context).size.height * 0.50,
                               child: new Scrollbar(
                                   child: new ListView.separated(
-                                itemCount: _paths != null && _paths.isNotEmpty ? _paths.length : 1,
+                                itemCount: _paths != null && _paths.isNotEmpty
+                                    ? _paths.length
+                                    : 1,
                                 itemBuilder: (BuildContext context, int index) {
-                                  final bool isMultiPath = _paths != null && _paths.isNotEmpty;
-                                  final String name = 'File $index: ' + (isMultiPath ? _paths.keys.toList()[index] : _fileName ?? '...');
-                                  final path = isMultiPath ? _paths.values.toList()[index].toString() : _path;
+                                  final bool isMultiPath =
+                                      _paths != null && _paths.isNotEmpty;
+                                  final String name = 'File $index: ' +
+                                      (isMultiPath
+                                          ? _paths.keys.toList()[index]
+                                          : _fileName ?? '...');
+                                  final path = isMultiPath
+                                      ? _paths.values.toList()[index].toString()
+                                      : _path;
 
                                   return new ListTile(
                                     title: new Text(
@@ -142,7 +165,9 @@ class _FilePickerDemoState extends State<FilePickerDemo> {
                                     subtitle: new Text(path),
                                   );
                                 },
-                                separatorBuilder: (BuildContext context, int index) => new Divider(),
+                                separatorBuilder:
+                                    (BuildContext context, int index) =>
+                                        new Divider(),
                               )),
                             )
                           : new Container(),